feat: prompt for project id if on interactive terminal
