+++ pkg/buildx/commands/build.go
@@ -623,9 +623,7 @@ func BuildCmd(dockerCli command.Cli) *cobra.Command {
 			}
 
 			options.project = helpers.ResolveProjectID(options.project, options.contextPath, options.dockerfileName)
-			if options.project == "" {
-				return errors.Errorf("unknown project ID (run `depot init` or use --project or $DEPOT_PROJECT_ID)")
-			}
+
 			buildPlatform, err := helpers.ResolveBuildPlatform(options.buildPlatform)
 			if err != nil {
 				return err
