+++ pkg/helpers/build.go
@@ -2,8 +2,10 @@ package helpers
 
 import (
 	"context"
+	"errors"
 	"os"
 
+	"github.com/bufbuild/connect-go"
 	depotbuild "github.com/depot/cli/pkg/build"
 	"github.com/depot/cli/pkg/profiler"
 	cliv1 "github.com/depot/cli/pkg/proto/depot/cli/v1"
@@ -19,6 +21,16 @@ func BeginBuild(ctx context.Context, req *cliv1.CreateBuildRequest, token string
 		build, err = depotbuild.NewBuild(ctx, req, token)
 	}
 	if err != nil {
+		var buildErr *connect.Error
+		if errors.As(err, &buildErr) && buildErr.Code() == connect.CodeNotFound {
+			selectedProject, err := InitializeProject(ctx, token, req.ProjectId)
+			if err != nil {
+				return depotbuild.Build{}, err
+			}
+
+			req.ProjectId = selectedProject.ID
+			return BeginBuild(ctx, req, token)
+		}
 		return depotbuild.Build{}, err
 	}
